backend/requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
djangorestframework==3.12.4
django-cors-headers==3.7.0